requirements.txt
absl-py==0.4.1
astor==0.7.1
astroid==2.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
gast==0.2.0
grpcio==1.15.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.5
protobuf==3.6.1
pylint==2.1.1
pymongo==3.7.1
PyYAML==3.13
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
